+++ backend/app/routes/api/plugins.py
@@ -6,7 +6,7 @@ from app.models.Bot import Bot
 from flask_jwt_extended import jwt_required
 from app.crud.bot import get_bot_options, set_bot_options
 from app.crud.plugins import new_plugin
-from app.crud.plugins import toggle_plugin, add_new_plagin, del_plagin, get_plugins
+from app.crud.plugins import toggle_plugin, add_new_plagin, del_plugin, get_plugins
 
 @api_bp.route("/plugins", methods=["GET"])
 @jwt_required()
@@ -17,7 +17,7 @@ def plagins():
 @api_bp.route("/plugins/<plugin_name>", methods=["DELETE"])
 @jwt_required()
 def delete_plugin(plugin_name: str):
-    return del_plagin(plugin_name)
+    return del_plugin(plugin_name)
 
 @api_bp.route("/plugins/<botname>/<int:id>/toggle", methods=["PUT"])
 @jwt_required()
@@ -35,6 +35,8 @@ def add_plugin(botname: str):
 @jwt_required()
 def get_plugins_options(botname: str, plugin_id: int):
     plugin = db.session.query(Plugin).filter(Plugin.id == plugin_id).first()
+    if not plugin:
+        return jsonify({"error": "Плагин не найден"}), 404
     name = plugin.name
     if not name:
         return jsonify({"error": "Название плагина не указано"}), 400
@@ -44,6 +46,9 @@ def get_plugins_options(botname: str, plugin_id: int):
 @jwt_required()
 def update_plugin_options(botname: str, plugin_id: int):
     plugin = db.session.query(Plugin).filter(Plugin.id == plugin_id).first()
+    if not plugin:
+        return jsonify({"error": "Плагин не найден"}), 404
+    
     name = plugin.name
     if not name:
         return jsonify({"error": "Название плагина не указано"}), 400
@@ -55,6 +60,9 @@ def update_plugin_options(botname: str, plugin_id: int):
 @jwt_required()
 def uninstall_plugin_bot(botname: str, plugin_name: str):
     bot = db.session.query(Bot).filter(Bot.name == botname).first()
+    if not bot:
+        return jsonify({"error": "Бот не найден"}), 404
+    
     for bp in bot.bot_plugins:
         if bp.plugin.name == plugin_name:
             db.session.delete(bp)
@@ -66,87 +74,7 @@ def uninstall_plugin_bot(botname: str, plugin_name: str):
 @jwt_required()
 def install_plugin():
     data = request.get_json()
-    if not data['name'] or not data['url']:
+    if not data['name'] or not data.get('download_link'):
         return jsonify({"error": "Название плагина не указано и/или URL"}), 400
     
     return new_plugin(data)
-
-
-pluginData = [
-    {
-        "id": 5,
-        "name": "Echo",
-        "enabled": False,
-        "description": "Простой плагин-эхо: повторяет сообщение пользователя.",
-        "price": 0,
-        "version": "1.0.0",
-        "author": "Aiser",
-        "license": "Ais",
-        "url": "https://aisblack.ru/static/plugins/Echo.zip",
-        "poster": "https://interesnyefakty.org/wp-content/uploads/chto-takoe-plagin.jpg"
-    },
-    {
-        "id": 6,
-        "name": "Welcome",
-        "enabled": False,
-        "description": "Отправляет приветственное сообщение новым пользователям.",
-        "price": 199,
-        "version": "1.2.0",
-        "author": "Aiser",
-        "license": "Ais",
-        "url": "https://aisblack.ru/static/plugins/Welcome.zip",
-        "poster": "https://content.timeweb.com/assets/65c70e62-4ae9-48bc-92ff-7886de5f50fa.jpg?width=3080&height=1600"
-    },
-    {
-        "id": 11,
-        "name": "Calculator",
-        "enabled": False,
-        "description": "Калькулятор.",
-        "price": 0,
-        "version": "1.0.2",
-        "author": "Aiser",
-        "license": "Ais",
-        "url": "https://aisblack.ru/static/plugins/Calculator.zip",
-        "poster": "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcT4msaMIrvPeFegjYbfOVjSWO5jEmnQQKKqvQ&s"
-    },
-    {
-        "id": 9,
-        "name": "Weather",
-        "enabled": False,
-        "description": "Показывает погоду.",
-        "price": 0,
-        "version": "1.0.2",
-        "author": "Aiser",
-        "license": "Ais",
-        "url": "https://aisblack.ru/static/plugins/Weather.zip",
-        "poster": "https://img.freepik.com/premium-vector/weather-logo-gradient-vector-icon-illustration_269830-2064.jpg"
-    },
-    {
-        "id": 10,
-        "name": "WebAppButton",
-        "enabled": False,
-        "description": "Веб-приложение Кнопка рядом с полем ввода.",
-        "price": 0,
-        "version": "1.0.2",
-        "author": "Aiser",
-        "license": "Ais",
-        "url": "https://aisblack.ru/static/plugins/WebAppButton.zip",
-        "poster": "https://habrastorage.org/getpro/habr/upload_files/159/0bf/7f2/1590bf7f2c9295f7934da9b760d8696e.jpeg"
-    },
-    {
-        "id": 7,
-        "name": "Translate",
-        "enabled": False,
-        "description": "Переводит сообщение в другую локаль.",
-        "price": 299,
-        "version": "1.0.0",
-        "author": "Aiser",
-        "license": "Ais",
-        "url": "https://aisblack.ru/static/plugins/Translate.zip",
-        "poster": "https://translations.telegram.org/img/translations/lang_banner.png?1"
-    }
-]
-
-@api_bp.route("/plugins/store", methods=["GET"])
-def get_plugins_store():
-    return jsonify(pluginData), 200
